💫CHANGED💫 Now using concurrency
Honestly this is just a large pile of AI slop, I'm glad it works!
I just do not care about this project and wish someone would make a pretty to do list that's developer friendly
Because this is needless effort. And I'm honestly just frustrated that it's twice as slow
As luajit :((